+++ packages/browser-wallet/CHANGELOG.md
@@ -11,6 +11,7 @@
 -   Verifiable credentials are now validated according to the schema when being added. This will e.g. block setting an attribute as an integer if the schema defines it as a string.
 -   Refreshed the schema for credential schemas so that attribute types are now restricted as expected (`string`, `integer` and the special types are allowed).
 -   An issue where credential schemas were not updated with the correct key.
+-   UI improvements to the credential selector.
 
 ## 1.1.5